create the scene for our flower garden
