+++ .vercel/project.json
@@ -1,0 +1,1 @@
+{"orgId":"team_GpVuPAGfhf031t7f8oTYDtMp","projectId":"prj_LUDWEMgDrczFT1fTqLIl5z5OtDGk"}
