+++ src/app/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 import Garden from '../garden/Garden';
 import Calendar from '../calendar/Calendar';
-import { PerspectiveCamera, Stars, OrbitControls, Sky, useTexture, Plane } from '@react-three/drei';
+import { PerspectiveCamera, OrbitControls, Sky } from '@react-three/drei';
 import { Canvas } from '@react-three/fiber';
 
 function App() {
